minor fixes

+++ scripts/customized_trainer.py
@@ -7,22 +7,32 @@ from transformers import (
     TrainerControl,
 )
 import os
-import re
-import torch
 from typing import Callable, Optional, Dict
 import shutil
 import json
 from transformers.trainer_utils import is_main_process
-from peft import PeftModel, PeftConfig
-from transformers import AutoModelForCausalLM, AutoTokenizer
-import logging, time, traceback
-from huggingface_hub.utils import HfHubHTTPError
-from huggingface_hub import HfApi
-import trainer.constants as train_cst
-import trainer.utils.training_paths as train_paths
+import wandb
+import torch
+
+
+MIS_MATCH_VOCAB_SIZE_MODELS = [
+    'NousResearch/Nous-Capybara-7B-V1',
+    'berkeley-nest/Starling-LM-7B-alpha',
+    'NousResearch/Hermes-2-Theta-Llama-3-8B',
+    'MNC-Jihun/Mistral-7B-AO-u0.5-b2-ver0.4'
+]
+
+ERROR_GENERATION_CONFIG_MODELS = [
+    "lmsys/vicuna-7b-v1.5", 
+    "lmsys/vicuna-13b-v1.5",
+    "NousResearch/Nous-Hermes-llama-2-7b", 
+    "defog/llama-3-sqlcoder-8b"
+]
 
 LOCAL_RANK = int(os.getenv("LOCAL_RANK", "0"))
 
+print(f"LOCAL_RANK: {LOCAL_RANK} in customized_trainer.py", flush=True)
+    
 class CustomEvalSaveCallback(TrainerCallback):
     def __init__(
         self,
@@ -30,6 +40,7 @@ class CustomEvalSaveCallback(TrainerCallback):
         submission_dir: str,
         output_dir: str,
         original_model_name: str,
+        max_steps: int = -1
     ):
         self.function_when_to_evaluate = function_when_to_evaluate
         self.submission_dir = submission_dir
@@ -38,32 +49,43 @@ class CustomEvalSaveCallback(TrainerCallback):
         self.update_best_checkpoint = False
         self.output_dir = output_dir
         self.original_model_name = original_model_name
+        self.max_steps = max_steps
+        self.has_checkpoint = False
+        self.save_only = False
         
     def compute_loss(self, state: TrainerState, metrics):
         return metrics.get("eval_loss", None)
 
     def on_step_end(self, args, state: TrainerState, control: TrainerControl, **kwargs):
         # Custom logic to decide whether to save or evaluate
+        # print(f"************* on_step_end: {state.global_step}, check eval", flush=True)
+        # TODO: implement the logic to save the model without evaluating if there is no check points --> avoid evaluating takes too much time
         when_to_eval = self.function_when_to_evaluate(state.global_step)
         if when_to_eval["eval"]:
-            print(f"Evaluating the model at step: {state.global_step} the reason: {when_to_eval['reason']}")
+            # do not allow the pod to be stopped by any reason 
+                # first check if there is at least one checkpoint or not 
+            print(f"Evaluating the model at step: {state.global_step} the reason: {when_to_eval['reason']}", flush=True)
             control.should_evaluate = True
-        if when_to_eval["save"]:
             control.should_save = True
-            control.should_training_stop = True
+            if when_to_eval["reason"] == "end_time":
+                if not self.has_checkpoint: # if there is no checkpoint, we just save the model, do not evaluate
+                    print(f"No checkpoint found, just save the model at step: {state.global_step}", flush=True)
+                    control.should_evaluate = False
+                    self.save_only = True
         return control
 
 
     def on_evaluate(
         self, args, state: TrainerState, control: TrainerControl, metrics, **kwargs
     ):
+        self.save_only = False
         # Append eval_loss to file
         eval_loss = self.compute_loss(state, metrics)
         if state.global_step < 2:
             return 
-        print(f"GO INTO CUSTOMIZED EVALUATE AT STEP: {state.global_step}")
+        print(f"GO INTO CUSTOMIZED EVALUATE AT STEP: {state.global_step}", flush=True)
         if self.best_checkpoint_info is None or eval_loss < self.best_checkpoint_info["loss"]:
-            print(f"Updating the best checkpoint info at step: {state.global_step} with eval_loss: {eval_loss}")
+            print(f"Updating the best checkpoint info at step: {state.global_step} with eval_loss: {eval_loss}", flush=True)
             self.best_checkpoint_info = {
                 "loss": eval_loss,
                 "step": state.global_step
@@ -71,55 +93,60 @@ class CustomEvalSaveCallback(TrainerCallback):
             self.update_best_checkpoint = True
         else:
             if self.best_checkpoint_info is not None:
-                print(f" At step: {state.global_step} The eval_loss: {eval_loss} is not smaller than the current best eval_loss: {self.best_checkpoint_info['loss']}, update_best_checkpoint={self.update_best_checkpoint}")
+                print(f" At step: {state.global_step} The eval_loss: {eval_loss} is not smaller than the current best eval_loss: {self.best_checkpoint_info['loss']}, update_best_checkpoint={self.update_best_checkpoint}", flush=True)
+            
 
     def on_save(self, args, state: TrainerState, control: TrainerControl, **kwargs):
-        if is_main_process(LOCAL_RANK):
-            print(f"Copy the latest checkpoint to the submission directory at step: {state.global_step}")
-
+        
+        if state.global_step == self.max_steps and self.max_steps != -1:
+            print(f"Stop training because of max steps: {self.max_steps}", flush=True)
+            control.should_training_stop = True
+        
+        self.has_checkpoint = True
+        
+        if not is_main_process(LOCAL_RANK): # if not main process, skip this
+            return 
+            
+        if self.save_only: # if only save, do not evaluate 
+            print(f"Only save the model at step: {state.global_step}, no evaluation", flush=True)
+            current_step = state.global_step
+            # Remove existing directory if it exists
             if os.path.exists(self.submission_dir):
                 shutil.rmtree(self.submission_dir)
-
-            checkpoint_dirs = [
-                d for d in os.listdir(self.output_dir)
-                if re.match(r'^checkpoint-\d+$', d)
-            ]
-            if not checkpoint_dirs:
-                print(f"[ERROR] No checkpoint directories found in {self.output_dir}. Skipping copy.")
-                return
-
-            latest_checkpoint = max(
-                checkpoint_dirs,
-                key=lambda x: int(x.split("-")[1])
+                
+            shutil.copytree(
+                os.path.join(self.output_dir, f"checkpoint-{current_step}"),
+                self.submission_dir
             )
-            checkpoint_path = os.path.join(self.output_dir, latest_checkpoint)
-
-            print(f"Latest checkpoint found: {latest_checkpoint}")
-            try:
-                print(f"Detected PEFT adapter: merging with base model...")
-                original_model_name = str(train_paths.get_text_base_model_path(self.original_model_name))
-                base_model = AutoModelForCausalLM.from_pretrained(original_model_name, trust_remote_code=True)
-
-                tokenizer = AutoTokenizer.from_pretrained(original_model_name, trust_remote_code=True)
-                tokenizer.save_pretrained(self.submission_dir)
-
-                num_tokens = len(tokenizer)
-                base_model.resize_token_embeddings(num_tokens)
-                peft_model = PeftModel.from_pretrained(base_model, checkpoint_path)
-                merged_model = peft_model.merge_and_unload()
-                merged_model.save_pretrained(self.submission_dir, safe_serialization=True, max_shard_size="4GB")
-
-                print(f"Merged full model saved to {self.submission_dir}")
-            except Exception as e:
-                print(f"Not a PEFT model or merge failed: {e}. Falling back to raw checkpoint copy.")
-                if os.path.exists(self.submission_dir):
-                    shutil.rmtree(self.submission_dir)
-                shutil.copytree(checkpoint_path, self.submission_dir)
-
-            patch_model_metadata(self.submission_dir, self.original_model_name)
-            patch_wandb_symlinks(train_cst.WANDB_LOGS_DIR)
-
             self.update_best_checkpoint = False
+            # add a loss.txt file to the submission directory
+            with open(os.path.join(self.submission_dir, "loss.txt"), "w") as f:
+                f.write(f"{current_step},no_eval")
+            
+            # release the flag
+            self.save_only = False
+            return 
+            
+        # Custom logic after model is saved
+        # You can trigger external services, logs, or backups here
+        if (
+            self.update_best_checkpoint
+            and is_main_process(LOCAL_RANK)
+        ):
+            print(f"Copy the best checkpoint to the submission directory at step: {state.global_step}", flush=True)
+            # Remove existing directory if it exists
+            if os.path.exists(self.submission_dir):
+                shutil.rmtree(self.submission_dir)
+            best_eval_loss = self.best_checkpoint_info["loss"]
+            shutil.copytree(
+                os.path.join(self.output_dir, f"checkpoint-{self.best_checkpoint_info['step']}"),
+                self.submission_dir
+            )
+            self.update_best_checkpoint = False
+            # add a loss.txt file to the submission directory
+            with open(os.path.join(self.submission_dir, "loss.txt"), "w") as f:
+                f.write(f"{self.best_checkpoint_info['step']},{best_eval_loss}")
+
 
 class GRPOCustomEvalSaveCallback(CustomEvalSaveCallback):
     def compute_loss(self, state: TrainerState, metrics):
@@ -127,9 +154,9 @@ class GRPOCustomEvalSaveCallback(CustomEvalSaveCallback):
         if state.log_history:
             last_log_entry = state.log_history[-1]
             eval_loss = last_log_entry.get("eval_reward", None)
-            print(f"choose eval_loss ({eval_loss}) as eval_reward from: last_log_entry: {last_log_entry}; \n metrics: {metrics}")
+            print(f"choose eval_loss ({eval_loss}) as eval_reward from: last_log_entry: {last_log_entry}; \n metrics: {metrics}", flush=True)
         else:
-            print(f"state.log_history is empty")
+            print(f"state.log_history is empty", flush=True)
             
         if eval_loss is not None:
             eval_loss = - eval_loss
@@ -150,89 +177,68 @@ def check_remaining_time_less_than_minutes(end_time: str, minutes: int) -> bool:
     time_diff = end_time - now
     result =  time_diff.total_seconds() < minutes * 60
     if result:
-        print(f"*** current time: {now} end_time: {end_time} time_diff: {time_diff}")
+        print(f"*** current time: {now} end_time: {end_time} time_diff: {time_diff}", flush=True)
     return result
 
 
 class WhenToEvalHandler:
-    def __init__(self, end_time: str, save_before_remaining_time: int = 3):
+    def __init__(self, end_time: str, save_before_remaining_time: int = 3, periodic_save_steps: int = -1, steps_per_epoch: int = -1, max_steps: int = -1):
         self.save_before_remaining_time = save_before_remaining_time
-        self.run_eval_save = False
+        self.run_eval = False
         self.end_time = end_time
+        self.periodic_save_steps = periodic_save_steps
+        self.steps_per_epoch = steps_per_epoch
+        self.max_steps = max_steps
 
     def __call__(self, global_step: int) -> dict:
-        if self.save_before_remaining_time > 0 and not self.run_eval_save:
+        
+        if self.steps_per_epoch != -1 and global_step % self.steps_per_epoch == 0 and global_step > 1:
+            return {"eval": True, "reason": "epoch"}
+        
+        if self.periodic_save_steps != -1 and global_step % self.periodic_save_steps == 0 and global_step > 1:
+            return {"eval": True, "reason": "periodic"}
+        
+        if self.save_before_remaining_time > 0 and not self.run_eval:
             if check_remaining_time_less_than_minutes(self.end_time, self.save_before_remaining_time):
+                print(f"***ALERT: The time is about to run out need to eval & save the model", flush=True)
                 # the eval time might be higher than the end_time, so we need to let the pod not stop by setting a flag for this
-                self.run_eval_save = True
-                return {"save": True, "eval": False, "reason": "end_time"}
+                self.run_eval = True
+                return {"eval": True, "reason": "end_time"}
+        
+        if self.max_steps != -1 and global_step == self.max_steps:
+            print(f"Stop training because of max steps: {self.max_steps}", flush=True)
+            return {"eval": True, "reason": "max_step"}
 
-        return {"save": False, "eval": False, "reason": "none"}
+        return {"eval": False, "reason": "none"}
 
 
-def patch_model_metadata(output_dir: str, base_model_id: str):
+def set_generation_config(model_name, model):
     try:
-        adapter_config_path = os.path.join(output_dir, "adapter_config.json")
-
-        if os.path.exists(adapter_config_path):
-            with open(adapter_config_path, "r") as f:
-                config = json.load(f)
-
-            config["base_model_name_or_path"] = base_model_id
-
-            with open(adapter_config_path, "w") as f:
-                json.dump(config, f, indent=2)
-
-            print(f"Updated adapter_config.json with base_model: {base_model_id}", flush=True)
-        else:
-            print(" adapter_config.json not found", flush=True)
-
-        readme_path = os.path.join(output_dir, "README.md")
-
-        if os.path.exists(readme_path):
-            with open(readme_path, "r") as f:
-                lines = f.readlines()
-
-            new_lines = []
-            for line in lines:
-                if line.strip().startswith("base_model:"):
-                    new_lines.append(f"base_model: {base_model_id}\n")
-                else:
-                    new_lines.append(line)
-
-            with open(readme_path, "w") as f:
-                f.writelines(new_lines)
-
-            print(f"Updated README.md with base_model: {base_model_id}", flush=True)
-        else:
-            print("README.md not found", flush=True)
-
-    except Exception as e:
-        print(f"Error updating metadata: {e}", flush=True)
+        if model_name in ERROR_GENERATION_CONFIG_MODELS:
+            model.generation_config = GenerationConfig(temperature=None, top_p=None)
+    except:
+        print(f"Error setting generation config for model {model_name}")
         pass
 
 
-def patch_wandb_symlinks(base_dir:str):
-    for root, _, files in os.walk(base_dir):
-        for name in files:
-            full_path = os.path.join(root, name)
+def resize_if_needed(model_name, model, token_nums):
+    try:
+        if model_name in MIS_MATCH_VOCAB_SIZE_MODELS:
+            model.resize_token_embeddings(token_nums)
+    except:
+        print(f"Error resizing token embeddings for model {model_name}")
+        pass
 
-            if os.path.islink(full_path):
-                target_path = os.readlink(full_path)
 
-                print(f"Symlink: {full_path} → {target_path}")
-                try:
-                    os.unlink(full_path)
-                except Exception as e:
-                    print(f"Failed to unlink {full_path}: {e}")
-                    continue
-
-                if os.path.exists(target_path):
-                    print("Copying real file")
-                    try:
-                        shutil.copy(target_path, full_path)
-                    except Exception as e:
-                        print(f"Failed to copy: {e}")
-                else:
-                    print("Target not found, creating dummy")
-                    pathlib.Path(full_path).touch()
+def init_wandb(train_request: Dict):
+    # set wandb_mode=offline; do not upload the data to wandb export WANDB_MODE=offline
+    return True
+    task_id = train_request["task_id"]
+    expected_repo_name = train_request["expected_repo_name"]
+    os.environ["WANDB_MODE"] = "offline"
+    os.environ["WANDB_DIR"] = train_request["wandb_log_dir"]
+    os.environ["WANDB_RUN_ID"] = f"{task_id}_{expected_repo_name}"
+    os.environ["WANDB_NAME"] = f"{task_id}_{expected_repo_name}"
+    if is_main_process(LOCAL_RANK):
+        os.makedirs(train_request["wandb_log_dir"], exist_ok=True)
+    return True
